desktop-autofill: Add skunkworks as preload codeowners
